+++ opentelemetry-api/src/opentelemetry/context/contextvars_context.py
@@ -12,14 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from contextvars import ContextVar
-from sys import version_info
 
 from opentelemetry.context.context import Context, _RuntimeContext
-
-if version_info < (3, 7):
-    import aiocontextvars  # type: ignore # pylint: disable=import-error
-
-    aiocontextvars  # pylint: disable=pointless-statement
 
 
 class ContextVarsRuntimeContext(_RuntimeContext):